Move currentyear handling out of guards
Read and update the currentyear variable, including year changes as
appropriate, including in cases where the end date is equal to the start
date or the end date is unknown. Also, update the current year whenever
a date is encountered whose year merely differs from the current year
(whether by being before or after it).